Bug#20561087 : REPLACE_USER_TABLE() DOES NOT CHECK ERROR WHEN READING FROM MYSQL.USER
Problem : Code responsible for reading/writing from/to acl tables
          assume that storage engine is always MYISAM. Hence, some
          errors reported by different storage engines (e.g. ndb)
          are not processed correctly.

Solution : Handle errors thrown by storage engine and bail out if
           error is not acceptable. This change also signals
           transaction rollback to storage engine in case of an
           unexpected error by storage engine.

+++ sql/auth/sql_user.cc
@@ -240,52 +240,15 @@ bool mysql_show_create_user(THD *thd, LEX_USER *user_name)
   Protocol *protocol= thd->get_protocol();
   USER_RESOURCES tmp_user_resource;
   enum SSL_type ssl_type;
-  char *ssl_cipher, *x509_issuer, *x509_subject, *ssl_info, *conn_attr;
+  char *ssl_cipher, *x509_issuer, *x509_subject;
   char buff[256];
   Item_string *field= NULL;
   List<Item> field_list;
   String sql_text(buff,sizeof(buff),system_charset_info);
-  TABLE_LIST tables;
-  TABLE *table;
-  uchar user_key[MAX_KEY_LENGTH];
   LEX_ALTER alter_info;
 
   DBUG_ENTER("mysql_show_create_user");
 
-  tables.init_one_table("mysql", 5, "user", 4, "user", TL_READ);
-  if (!(table= open_ltable(thd, &tables, TL_READ, MYSQL_LOCK_IGNORE_TIMEOUT)))
-    DBUG_RETURN(1);
-
-  if (!table->key_info)
-  {
-    my_error(ER_TABLE_CORRUPT, MYF(0), table->s->db.str,
-             table->s->table_name.str);
-    DBUG_RETURN(1);
-  }
-  table->use_all_columns();
-  table->field[MYSQL_USER_FIELD_HOST]->store(user_name->host.str,
-                                             user_name->host.length,
-                                             system_charset_info);
-  table->field[MYSQL_USER_FIELD_USER]->store(user_name->user.str,
-                                             user_name->user.length,
-                                             system_charset_info);
-  key_copy(user_key, table->record[0], table->key_info,
-           table->key_info->key_length);
-
-  if (table->file->ha_index_read_idx_map(table->record[0], 0, user_key,
-                                         HA_WHOLE_KEY,
-                                         HA_READ_KEY_EXACT))
-  {
-    String wrong_users;
-    append_user(thd, &wrong_users, user_name, wrong_users.length() > 0, false);
-    /* if user does not exists report error */
-    my_error(ER_CANNOT_USER, MYF(0), "SHOW CREATE USER",
-             wrong_users.c_ptr_safe());
-    close_acl_tables(thd);
-    DBUG_RETURN(1);
-  }
-
-  store_record(table,record[1]);
   mysql_mutex_lock(&acl_cache->lock);
   if (!(acl_user= find_acl_user(user_name->host.str, user_name->user.str, TRUE)))
   {
@@ -294,7 +257,6 @@ bool mysql_show_create_user(THD *thd, LEX_USER *user_name)
     append_user(thd, &wrong_users, user_name, wrong_users.length() > 0, false);
     my_error(ER_CANNOT_USER, MYF(0), "SHOW CREATE USER",
              wrong_users.c_ptr_safe());
-    close_acl_tables(thd);
     DBUG_RETURN(1);
   }
   /* fill in plugin, auth_str from acl_user */
@@ -306,63 +268,19 @@ bool mysql_show_create_user(THD *thd, LEX_USER *user_name)
   user_name->uses_identified_by_password_clause= false;
   user_name->uses_authentication_string_clause= false;
 
-  /* read user resources, ssl attributes from table */
-  ssl_info= get_field(thd->mem_root, table->field[MYSQL_USER_FIELD_SSL_TYPE]);
-  if (!ssl_info)
-    acl_user->ssl_type=SSL_TYPE_NONE;
-  else if (!strcmp(ssl_info, "ANY"))
-    acl_user->ssl_type=SSL_TYPE_ANY;
-  else if (!strcmp(ssl_info, "X509"))
-    acl_user->ssl_type=SSL_TYPE_X509;
-  else
-    acl_user->ssl_type=SSL_TYPE_SPECIFIED;
-
-  if (table->s->fields > MYSQL_USER_FIELD_SSL_CIPHER)
-    acl_user->ssl_cipher= get_field(&global_acl_memory,
-                                table->field[MYSQL_USER_FIELD_SSL_CIPHER]);
-  if (table->s->fields > MYSQL_USER_FIELD_X509_ISSUER)
-    acl_user->x509_issuer= get_field(&global_acl_memory,
-                               table->field[MYSQL_USER_FIELD_X509_ISSUER]);
-  if (table->s->fields > MYSQL_USER_FIELD_X509_SUBJECT)
-    acl_user->x509_subject= get_field(&global_acl_memory,
-                               table->field[MYSQL_USER_FIELD_X509_SUBJECT]);
-
-  if (table->s->fields > MYSQL_USER_FIELD_MAX_USER_CONNECTIONS)
-  {
-     conn_attr= get_field(thd->mem_root,
-                          table->field[MYSQL_USER_FIELD_MAX_USER_CONNECTIONS]);
-     acl_user->user_resource.user_conn= conn_attr ? atoi(conn_attr) : 0;
-     if (acl_user->user_resource.user_conn)
-       acl_user->user_resource.specified_limits|= USER_RESOURCES::USER_CONNECTIONS;
-  }
-  if (table->s->fields > MYSQL_USER_FIELD_MAX_QUESTIONS)
-  {
-     conn_attr= get_field(thd->mem_root,
-                          table->field[MYSQL_USER_FIELD_MAX_QUESTIONS]);
-     acl_user->user_resource.questions= conn_attr ? atoi(conn_attr) : 0;
-     if (acl_user->user_resource.questions)
-       acl_user->user_resource.specified_limits|= USER_RESOURCES::QUERIES_PER_HOUR;
-  }
-  if (table->s->fields > MYSQL_USER_FIELD_MAX_UPDATES)
-  {
-     conn_attr= get_field(thd->mem_root,
-                          table->field[MYSQL_USER_FIELD_MAX_UPDATES]);
-     acl_user->user_resource.updates= conn_attr ? atoi(conn_attr) : 0;
-     if (acl_user->user_resource.updates)
-       acl_user->user_resource.specified_limits|= USER_RESOURCES::UPDATES_PER_HOUR;
-  }
-  if (table->s->fields > MYSQL_USER_FIELD_MAX_CONNECTIONS)
-  {
-     conn_attr= get_field(thd->mem_root,
-                          table->field[MYSQL_USER_FIELD_MAX_CONNECTIONS]);
-     acl_user->user_resource.conn_per_hour= conn_attr ? atoi(conn_attr) : 0;
-     if (acl_user->user_resource.conn_per_hour)
-       acl_user->user_resource.specified_limits|= USER_RESOURCES::CONNECTIONS_PER_HOUR;
-  }
-
   /* make a copy of user resources, ssl and password expire attributes */
   tmp_user_resource= lex->mqh;
   lex->mqh= acl_user->user_resource;
+
+  /* Set specified_limits flags so user resources are shown properly. */
+  if (lex->mqh.user_conn)
+    lex->mqh.specified_limits|= USER_RESOURCES::USER_CONNECTIONS;
+  if (lex->mqh.questions)
+    lex->mqh.specified_limits|= USER_RESOURCES::QUERIES_PER_HOUR;
+  if (lex->mqh.updates)
+    lex->mqh.specified_limits|= USER_RESOURCES::UPDATES_PER_HOUR;
+  if (lex->mqh.conn_per_hour)
+    lex->mqh.specified_limits|= USER_RESOURCES::CONNECTIONS_PER_HOUR;
 
   ssl_type= lex->ssl_type;
   ssl_cipher= lex->ssl_cipher;
@@ -418,7 +336,6 @@ err:
   lex->alter_password= alter_info;
 
   mysql_mutex_unlock(&acl_cache->lock);
-  close_acl_tables(thd);
   my_eof(thd);
   DBUG_RETURN(error);
 }
@@ -692,7 +609,8 @@ bool change_password(THD *thd, const char *host, const char *user,
   ulong what_to_set= 0;
   bool save_binlog_row_based;
   size_t new_password_len= strlen(new_password);
-  bool result= 1;
+  bool result= true, rollback_whole_statement= false;
+  int ret;
 
   DBUG_ENTER("change_password");
   DBUG_PRINT("enter",("host: '%s'  user: '%s'  new_password: '%s'",
@@ -794,10 +712,13 @@ bool change_password(THD *thd, const char *host, const char *user,
     mysql_mutex_unlock(&acl_cache->lock);
     goto end;
   }
-  if (replace_user_table(thd, table, combo, 0, 0, 1, what_to_set))
+  ret= replace_user_table(thd, table, combo, 0, false, true, what_to_set);
+  if (ret)
   {
-    result= 1;
     mysql_mutex_unlock(&acl_cache->lock);
+    result= 1;
+    if (ret < 0)
+      rollback_whole_statement= true;
     goto end;
   }
   if (!update_sctx_cache(thd->security_context(), acl_user, false) &&
@@ -833,7 +754,9 @@ bool change_password(THD *thd, const char *host, const char *user,
   result= write_bin_log(thd, true, buff, query_length,
                         table->file->has_transactions());
 end:
-  result|= acl_trans_commit_and_close_tables(thd);
+  result|= acl_end_trans_and_close_tables(thd,
+                                          thd->transaction_rollback_request ||
+                                          rollback_whole_statement);
 
   if (!result)
     acl_notify_htons(thd, buff, query_length);
@@ -1142,7 +1065,7 @@ static int handle_grant_data(TABLE_LIST *tables, bool drop,
   if ((found= handle_grant_table(tables, 0, drop, user_from, user_to)) < 0)
   {
     /* Handle of table failed, don't touch the in-memory array. */
-    result= -1;
+    DBUG_RETURN(-1);
   }
   else
   {
@@ -1166,7 +1089,7 @@ static int handle_grant_data(TABLE_LIST *tables, bool drop,
   if ((found= handle_grant_table(tables, 1, drop, user_from, user_to)) < 0)
   {
     /* Handle of table failed, don't touch the in-memory array. */
-    result= -1;
+    DBUG_RETURN(-1);
   }
   else
   {
@@ -1190,7 +1113,7 @@ static int handle_grant_data(TABLE_LIST *tables, bool drop,
   if ((found= handle_grant_table(tables, 4, drop, user_from, user_to)) < 0)
   {
     /* Handle of table failed, don't touch in-memory array. */
-    result= -1;
+    DBUG_RETURN(-1);
   }
   else
   {
@@ -1230,7 +1153,7 @@ static int handle_grant_data(TABLE_LIST *tables, bool drop,
   if ((found= handle_grant_table(tables, 2, drop, user_from, user_to)) < 0)
   {
     /* Handle of table failed, don't touch columns and in-memory array. */
-    result= -1;
+    DBUG_RETURN(-1);
   }
   else
   {
@@ -1246,7 +1169,7 @@ static int handle_grant_data(TABLE_LIST *tables, bool drop,
     if ((found= handle_grant_table(tables, 3, drop, user_from, user_to)) < 0)
     {
       /* Handle of table failed, don't touch the in-memory array. */
-      result= -1;
+      DBUG_RETURN(-1);
     }
     else
     {
@@ -1266,7 +1189,7 @@ static int handle_grant_data(TABLE_LIST *tables, bool drop,
     if ((found= handle_grant_table(tables, 5, drop, user_from, user_to)) < 0)
     {
       /* Handle of table failed, don't touch the in-memory array. */
-      result= -1;
+      DBUG_RETURN(-1);
     }
     else
     {
@@ -1308,7 +1231,7 @@ bool mysql_create_user(THD *thd, List <LEX_USER> &list)
   bool transactional_tables;
   ulong what_to_update= 0;
   bool is_anonymous_user= false;
-
+  bool rollback_whole_statement= false;
   DBUG_ENTER("mysql_create_user");
 
   /*
@@ -1362,20 +1285,23 @@ bool mysql_create_user(THD *thd, List <LEX_USER> &list)
       Search all in-memory structures and grant tables
       for a mention of the new user name.
     */
-    if (handle_grant_data(tables, 0, user_name, NULL))
+    int ret1= 0, ret2= 0;
+    if ((ret1= handle_grant_data(tables, 0, user_name, NULL)) ||
+        (ret2= replace_user_table(thd, tables[0].table, user_name, 0,
+                                  false, true, what_to_update)))
     {
-      append_user(thd, &wrong_users, user_name, wrong_users.length() > 0,
-                  false);
-      result= TRUE;
-      continue;
-    }
-
-    if (replace_user_table(thd, tables[0].table, user_name, 0, 0, 1, what_to_update))
-    {
-      append_user(thd, &wrong_users, user_name, wrong_users.length() > 0,
-                  false);
-      result= TRUE;
-      continue;
+      result= true;
+      if (ret1 < 0 || ret2 < 0)
+      {
+        rollback_whole_statement= true;
+        break;
+      }
+      else
+      {
+        append_user(thd, &wrong_users, user_name, wrong_users.length() > 0,
+                    false);
+        continue;
+      }
     }
 
     some_users_created= TRUE;
@@ -1383,7 +1309,7 @@ bool mysql_create_user(THD *thd, List <LEX_USER> &list)
 
   mysql_mutex_unlock(&acl_cache->lock);
 
-  if (result)
+  if (result && !rollback_whole_statement)
   {
     if (is_anonymous_user)
       my_error(ER_CANNOT_USER, MYF(0), "CREATE USER", "anonymous user");
@@ -1409,7 +1335,9 @@ bool mysql_create_user(THD *thd, List <LEX_USER> &list)
 
   lock.unlock();
 
-  result|= acl_trans_commit_and_close_tables(thd);
+  result|= acl_end_trans_and_close_tables(thd,
+                                          thd->transaction_rollback_request ||
+                                          rollback_whole_statement);
 
   if (some_users_created && !result)
     acl_notify_htons(thd, thd->query().str, thd->query().length);
@@ -1446,6 +1374,7 @@ bool mysql_drop_user(THD *thd, List <LEX_USER> &list)
   sql_mode_t old_sql_mode= thd->variables.sql_mode;
   bool save_binlog_row_based;
   bool transactional_tables;
+  bool rollback_whole_statement= false;
   DBUG_ENTER("mysql_drop_user");
 
   /*
@@ -1478,13 +1407,20 @@ bool mysql_drop_user(THD *thd, List <LEX_USER> &list)
       result= TRUE;
       continue;
     }  
-    if (handle_grant_data(tables, 1, user_name, NULL) <= 0)
+    int ret= handle_grant_data(tables, 1, user_name, NULL);
+    if (ret <= 0)
     {
+      result= true;
+
+      if (ret < 0)
+      {
+        rollback_whole_statement= true;
+        break;
+      }
       append_user(thd, &wrong_users, user_name, wrong_users.length() > 0, FALSE);
-      result= TRUE;
-      continue;
     }
-    some_users_deleted= TRUE;
+    else
+      some_users_deleted= true;
   }
 
   /* Rebuild 'acl_check_hosts' since 'acl_users' has been modified */
@@ -1492,7 +1428,7 @@ bool mysql_drop_user(THD *thd, List <LEX_USER> &list)
 
   mysql_mutex_unlock(&acl_cache->lock);
 
-  if (result)
+  if (result && !rollback_whole_statement)
     my_error(ER_CANNOT_USER, MYF(0), "DROP USER", wrong_users.c_ptr_safe());
 
   if (some_users_deleted)
@@ -1501,7 +1437,10 @@ bool mysql_drop_user(THD *thd, List <LEX_USER> &list)
 
   lock.unlock();
 
-  result|= acl_trans_commit_and_close_tables(thd);
+  result|=
+    acl_end_trans_and_close_tables(thd,
+                                   thd->transaction_rollback_request ||
+                                   rollback_whole_statement);
 
   if (some_users_deleted && !result)
     acl_notify_htons(thd, thd->query().str, thd->query().length);
@@ -1539,6 +1478,7 @@ bool mysql_rename_user(THD *thd, List <LEX_USER> &list)
   bool some_users_renamed= FALSE;
   bool save_binlog_row_based;
   bool transactional_tables;
+  bool rollback_whole_statement= false;
   DBUG_ENTER("mysql_rename_user");
 
   /*
@@ -1581,11 +1521,36 @@ bool mysql_rename_user(THD *thd, List <LEX_USER> &list)
       Search all in-memory structures and grant tables
       for a mention of the new user name.
     */
-    if (handle_grant_data(tables, 0, user_to, NULL) ||
-        handle_grant_data(tables, 0, user_from, user_to) <= 0)
+    int ret= handle_grant_data(tables, 0, user_to, NULL);
+
+    if (ret != 0)
     {
+      result= true;
+
+      if (ret < 0)
+      {
+        rollback_whole_statement= true;
+        break;
+      }
+
+      append_user(thd, &wrong_users, user_from, wrong_users.length() > 0,
+                  false);
+      continue;
+    }
+
+    ret= handle_grant_data(tables, 0, user_from, user_to);
+
+    if (ret <= 0)
+    {
+      result= true;
+
+      if (ret < 0)
+      {
+        rollback_whole_statement= true;
+        break;
+      }
+
       append_user(thd, &wrong_users, user_from, wrong_users.length() > 0, FALSE);
-      result= TRUE;
       continue;
     }
     some_users_renamed= TRUE;
@@ -1596,7 +1561,7 @@ bool mysql_rename_user(THD *thd, List <LEX_USER> &list)
 
   mysql_mutex_unlock(&acl_cache->lock);
 
-  if (result)
+  if (result && !rollback_whole_statement)
     my_error(ER_CANNOT_USER, MYF(0), "RENAME USER", wrong_users.c_ptr_safe());
   
   if (some_users_renamed)
@@ -1605,7 +1570,10 @@ bool mysql_rename_user(THD *thd, List <LEX_USER> &list)
 
   lock.unlock();
 
-  result|= acl_trans_commit_and_close_tables(thd);
+  result|=
+    acl_end_trans_and_close_tables(thd,
+                                   thd->transaction_rollback_request ||
+                                   rollback_whole_statement);
 
   if (some_users_renamed && !result)
     acl_notify_htons(thd, thd->query().str, thd->query().length);
@@ -1643,6 +1611,7 @@ bool mysql_alter_user(THD *thd, List <LEX_USER> &list)
   bool some_user_altered= false;
   bool save_binlog_row_based;
   bool is_privileged_user= false;
+  bool rollback_whole_statement= false;
 
   DBUG_ENTER("mysql_alter_user");
 
@@ -1751,12 +1720,18 @@ bool mysql_alter_user(THD *thd, List <LEX_USER> &list)
     }
 
     /* update the mysql.user table */
-    if (replace_user_table(thd, table, user_from, 0, 0, 1,
-                           what_to_alter))
+    int ret= replace_user_table(thd, table, user_from, 0, false, true,
+                                what_to_alter);
+    if (ret)
     {
+      result= true;
+      if (ret < 0)
+      {
+        rollback_whole_statement= true;
+        break;
+      }
       append_user(thd, &wrong_users, user_from, wrong_users.length() > 0,
                   false);
-      result= TRUE;
       continue;
     }
     some_user_altered= true;
@@ -1767,7 +1742,7 @@ bool mysql_alter_user(THD *thd, List <LEX_USER> &list)
   acl_cache->clear(1);                          // Clear locked hostname cache
   mysql_mutex_unlock(&acl_cache->lock);
 
-  if (result)
+  if (result && !rollback_whole_statement)
   {
     if (is_anonymous_user)
       my_error(ER_PASSWORD_EXPIRE_ANONYMOUS_USER, MYF(0));
@@ -1790,7 +1765,10 @@ bool mysql_alter_user(THD *thd, List <LEX_USER> &list)
 
   lock.unlock();
 
-  result|= acl_trans_commit_and_close_tables(thd);
+  result|=
+    acl_end_trans_and_close_tables(thd,
+                                   thd->transaction_rollback_request ||
+                                   rollback_whole_statement);
 
   if (some_user_altered && !result)
     acl_notify_htons(thd, thd->query().str, thd->query().length);
